Conflicts:

CoffeeMachine.xcodeproj/project.xcworkspace/xcuserdata/dancho.xcuserdata
d/UserInterfaceState.xcuserstate

CoffeeMachine.xcodeproj/xcuserdata/dancho.xcuserdatad/xcdebugger/Breakpo
ints.xcbkptlist

+++ CoffeeMachine/Model/Drink.h
@@ -11,7 +11,7 @@
 @interface Drink : NSObject <NSCopying, NSCoding>
 
 @property (strong) NSString *name;
-@property NSUInteger price;
+@property int price;
 -(NSString*)description;
 - (BOOL)isEqual:(Drink*)other;
 - (NSUInteger)hash;
